Use storyboard and FactoryMethodInjectable
