+++ Example/Example/View/Detail/Detail.storyboard
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="13196" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="MLa-Tf-otS">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13173"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Detail View Controller-->
+        <scene sceneID="z11-Oh-c2e">
+            <objects>
+                <viewController id="MLa-Tf-otS" customClass="DetailViewController" customModule="Example" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="Jti-Od-JST">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" translatesAutoresizingMaskIntoConstraints="NO" id="YTW-ff-Ypg">
+                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                                <subviews>
+                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="jAa-PB-Bs0">
+                                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                                        <subviews>
+                                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="LR7-yi-7I8">
+                                                <rect key="frame" x="166.5" y="323" width="42" height="21"/>
+                                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                                <nil key="textColor"/>
+                                                <nil key="highlightedColor"/>
+                                            </label>
+                                        </subviews>
+                                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                                        <constraints>
+                                            <constraint firstItem="LR7-yi-7I8" firstAttribute="centerY" secondItem="jAa-PB-Bs0" secondAttribute="centerY" id="71D-Vg-7fb"/>
+                                            <constraint firstItem="LR7-yi-7I8" firstAttribute="centerX" secondItem="jAa-PB-Bs0" secondAttribute="centerX" id="JvN-PR-owL"/>
+                                        </constraints>
+                                    </view>
+                                </subviews>
+                                <constraints>
+                                    <constraint firstItem="jAa-PB-Bs0" firstAttribute="width" secondItem="YTW-ff-Ypg" secondAttribute="width" id="90Y-U8-wIw"/>
+                                    <constraint firstAttribute="trailing" secondItem="jAa-PB-Bs0" secondAttribute="trailing" id="Snb-Hy-dTk"/>
+                                    <constraint firstAttribute="bottom" secondItem="jAa-PB-Bs0" secondAttribute="bottom" id="XUb-5K-BuG"/>
+                                    <constraint firstItem="jAa-PB-Bs0" firstAttribute="leading" secondItem="YTW-ff-Ypg" secondAttribute="leading" id="YGN-5F-7Js"/>
+                                    <constraint firstItem="jAa-PB-Bs0" firstAttribute="height" secondItem="YTW-ff-Ypg" secondAttribute="height" id="rPK-Fo-2MR"/>
+                                    <constraint firstItem="jAa-PB-Bs0" firstAttribute="top" secondItem="YTW-ff-Ypg" secondAttribute="top" id="x0k-Zi-FcG"/>
+                                </constraints>
+                            </scrollView>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <constraints>
+                            <constraint firstItem="huy-5K-ObI" firstAttribute="top" secondItem="YTW-ff-Ypg" secondAttribute="top" constant="20" id="GUd-tD-hNX"/>
+                            <constraint firstItem="YTW-ff-Ypg" firstAttribute="leading" secondItem="huy-5K-ObI" secondAttribute="leading" id="WCn-ME-6nn"/>
+                            <constraint firstItem="huy-5K-ObI" firstAttribute="trailing" secondItem="YTW-ff-Ypg" secondAttribute="trailing" id="dgv-86-UlG"/>
+                            <constraint firstItem="huy-5K-ObI" firstAttribute="bottom" secondItem="YTW-ff-Ypg" secondAttribute="bottom" id="lvJ-1p-qgf"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="huy-5K-ObI"/>
+                    </view>
+                    <connections>
+                        <outlet property="label" destination="LR7-yi-7I8" id="QL8-ZR-XD7"/>
+                    </connections>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="UMz-UO-Ft0" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="173.59999999999999" y="145.27736131934034"/>
+        </scene>
+    </scenes>
+</document>
